Add debug defines to common.hpp, release point for code.
Added #defines to turn on/off the debug prints

Reduced the size of Dataset so it can run under qemu
This will allow it to get past init() of memory

Current state of code:
--verify works with 1 thread, multiple nonces, different seeds
  - Current output from --verify does not match expected output hardcoded
	into code.
  - Setting multiple threads crashes on NULL ptr access

--mine crashes on randomx_init_dataset(). The cache ptr is being overwritten
in the code.  This causes an exception on first access to cache ptr
after randomx_init_dataset() call. Debug output is currently turned on
for this.

Interpreted mode works with 1 thread, crashes with NULL ptr access with threads > 1