Fixed a few more corner cases..
Check is in place for all 32bit constants for loading
into a register.

Still fails and returns same data on all passes for
Dataset initialization (compiler) test.

Problem is in superscalar generated code. If this is commented out
all 4 passes return unique data.

Hashes are still not correct, above issue with dataset tests
is related.