Update to IROR_C()
No longer has it's own braces for the code.

If IROR_C() is commented out data is random on each
pass of initDatasetInit().  Data is NOT consistent from
one run of randomx-tests to another.

If IROR_C() is not commented out, the data is the same for
each pass of initDatasetInit()